replace datadog_gauge
