+++ corehq/motech/repeaters/tasks.py
@@ -6,6 +6,7 @@ from celery.schedules import crontab
 from celery.task import periodic_task, task
 from celery.utils.log import get_task_logger
 
+from corehq.util.metrics.metrics import metrics_gauge_task
 from dimagi.utils.couch import get_redis_lock
 from dimagi.utils.couch.undo import DELETED_SUFFIX
 
@@ -25,7 +26,6 @@ from corehq.privileges import DATA_FORWARDING, ZAPIER_INTEGRATION
 from corehq.util.datadog.gauges import (
     datadog_bucket_timer,
     datadog_counter,
-    datadog_gauge_task,
 )
 from corehq.util.datadog.utils import make_buckets_from_timedeltas
 from corehq.util.soft_assert import soft_assert
@@ -137,7 +137,7 @@ def process_repeat_record(repeat_record):
         logging.exception('Failed to process repeat record: {}'.format(repeat_record._id))
 
 
-repeaters_overdue = datadog_gauge_task(
+repeaters_overdue = metrics_gauge_task(
     'commcare.repeaters.overdue',
     get_overdue_repeat_record_count,
     run_every=crontab()  # every minute
