+++ corehq/util/datadog/gauges.py
@@ -9,23 +9,6 @@ from corehq.util.datadog.utils import bucket_value
 from corehq.util.timer import TimingContext
 
 
-def datadog_gauge_task(name, fn, run_every, enforce_prefix='commcare'):
-    """
-    helper for easily registering datadog gauges to run periodically
-
-    To update a datadog gauge on a schedule based on the result of a function
-    just add to your app's tasks.py:
-
-        my_calculation = datadog_gauge_task('my.datadog.metric', my_calculation_function,
-                                            run_every=crontab(minute=0))
-
-    """
-    _enforce_prefix(name, enforce_prefix)
-
-    datadog_gauge = _DatadogGauge(name, fn, run_every)
-    return datadog_gauge.periodic_task()
-
-
 def datadog_histogram(name, value, enforce_prefix='commcare', tags=None):
     """
     Usage: Used to track the statistical distribution of a set of values over a statsd flush period.
@@ -98,23 +81,6 @@ def datadog_bucket_timer(metric, tags, timing_buckets, callback=None):
 
     timer.stop = new_stop
     return timer
-
-
-class _DatadogGauge(object):
-
-    def __init__(self, name, fn, run_every):
-        self.name = name
-        self.fn = fn
-        self.run_every = run_every
-
-    def periodic_task(self):
-        @periodic_task(serializer='pickle', queue='background_queue', run_every=self.run_every,
-                       acks_late=True, ignore_result=True)
-        @wraps(self.fn)
-        def inner(*args, **kwargs):
-            statsd.gauge(self.name, self.fn(*args, **kwargs))
-
-        return inner
 
 
 def _enforce_prefix(name, prefix):
